Fix some STM32 ethernet stuff that was not linking correctly
