+++ connectivity/drivers/emac/TARGET_STM/CMakeLists.txt
@@ -19,5 +19,4 @@ target_include_directories(mbed-emac
 target_sources(mbed-emac
     PRIVATE
         stm32xx_emac.cpp
-        stm32xx_eth_irq_callback.cpp
 )
