+++ connectivity/lwipstack/CMakeLists.txt
@@ -102,3 +102,4 @@ target_link_libraries(mbed-lwipstack
     PRIVATE
         mbed-rtos-flags
 )
+
